DriverLog: Rename to uppercase.
